+++ packages/block-library/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Enhancements
+
+-   Block Library: Restore the optimization that prevents experimental block code from being bundled in WordPress core builds by adding an ESBuild plugin ([#79203](https://github.com/WordPress/gutenberg/issues/79203)).
+
 ### Breaking Changes
 
 -   Removed the `@wordpress/block-library/babel-plugin` export. It was an internal transform for stripping experimental blocks from WordPress core builds and is no longer used by Gutenberg's build process ([#79162](https://github.com/WordPress/gutenberg/pull/79162)).